backend/pyproject.toml
[project]
name = "autopromo-backend"
version = "0.1.0"
description = "AutoPromo Cloud - HFT SaaS for Affiliate Marketing"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.3",
    "asyncpg>=0.29.0",
    "redis>=5.2.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.7.4",
]
